Fixing Android errors in merge conflict

+++ Frontend/NovelloApp/app/src/main/java/com/yn_1/novello_app/book/Book.java
@@ -10,7 +10,7 @@ public class Book {
     String author;
     int publicationYear;
     String isbn;
-    int rating;
+    double rating;
     double price;
     String description;
     private String readingURL;
